requirements.txt
Django>=4.0
djangorestframework
psycopg2-binary
openpyxl
celery
redis
pandas
